+++ africanus/model/wsclean/__init__.py
@@ -1,0 +1,4 @@
+__all__ = ["load", "spectra"]
+
+from africanus.model.wsclean.file_model import load
+from africanus.model.wsclean.spec_model import spectra