+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 * Support `complete` and `complete-cuda` to support non-GPU installs (:pr:`87`)
 * Gaussian Shape Parameter Implementation (:pr:`82`, :pr:`83`)
+* WSClean Spectral Model (:pr:`81`)
 * Compare predict versus MeqTrees (:pr:`79`)
 * Time and channel averaging (:pr:`75`)
 * cupy implementation of `predict_vis` (:pr:`73`)